# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  agent: "agent",
  admin: "admin",
} as const;

export interface UserProfile {
  id: string;
  telegramId: string;
  username: string;
  balanceUsd: number;
  balanceSyp: number;
  role: UserProfileRole;
}

export interface Category {
  id: string;
  name: string;
  image: string;
  order: number;
  active: boolean;
  productCount: number;
}

export type ProductProductType =
  (typeof ProductProductType)[keyof typeof ProductProductType];

export const ProductProductType = {
  amount: "amount",
  package: "package",
} as const;

export interface Product {
  id: string;
  name: string;
  categoryId: string;
  categoryName: string;
  image: string;
  priceUsd: number;
  priceSyp: number;
  productType: ProductProductType;
  available: boolean;
  minQty?: number;
  maxQty?: number;
  description?: string;
  featured: boolean;
}

export type NewsItemType = (typeof NewsItemType)[keyof typeof NewsItemType];

export const NewsItemType = {
  general: "general",
  offer: "offer",
  alert: "alert",
  new_service: "new_service",
} as const;

export interface NewsItem {
  id: string;
  content: string;
  type: NewsItemType;
}

export interface Banner {
  id: string;
  image: string;
  title: string;
  link?: string;
}

export type PaymentMethodCode =
  (typeof PaymentMethodCode)[keyof typeof PaymentMethodCode];

export const PaymentMethodCode = {
  sham_cash: "sham_cash",
  binance_pay: "binance_pay",
  syriatel_cash: "syriatel_cash",
  mtn_cash: "mtn_cash",
  usdt_auto: "usdt_auto",
} as const;

export interface PaymentMethod {
  id: string;
  code: PaymentMethodCode;
  name: string;
  subtitle: string;
  instructions?: string;
  walletAddress?: string;
  qrImage?: string;
  minAmount: number;
  active: boolean;
}

export interface SocialLink {
  id: string;
  platform: string;
  url: string;
  label: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  wait: "wait",
  accept: "accept",
  reject: "reject",
} as const;

export interface Order {
  id: string;
  orderNumber: string;
  productId: string;
  productName: string;
  productImage?: string;
  quantity: number;
  userIdentifier?: string;
  totalUsd: number;
  totalSyp: number;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateOrderBody {
  productId: string;
  quantity: number;
  userIdentifier?: string;
}

export interface OrdersSummary {
  totalAcceptedUsd: number;
  waitCount: number;
  acceptCount: number;
  totalCount: number;
}

export type DepositCurrency =
  (typeof DepositCurrency)[keyof typeof DepositCurrency];

export const DepositCurrency = {
  USD: "USD",
  SYP: "SYP",
} as const;

export type DepositStatus = (typeof DepositStatus)[keyof typeof DepositStatus];

export const DepositStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Deposit {
  id: string;
  amountUsd: number;
  amountSyp?: number;
  currency: DepositCurrency;
  method: string;
  methodLabel: string;
  transactionId: string;
  status: DepositStatus;
  createdAt: string;
}

export type CreateDepositBodyCurrency =
  (typeof CreateDepositBodyCurrency)[keyof typeof CreateDepositBodyCurrency];

export const CreateDepositBodyCurrency = {
  USD: "USD",
  SYP: "SYP",
} as const;

export interface CreateDepositBody {
  method: string;
  currency: CreateDepositBodyCurrency;
  amount: number;
  transactionId: string;
}

export interface DepositsSummary {
  totalApprovedUsd: number;
  pendingCount: number;
  approvedCount: number;
  totalCount: number;
}

export type ListProductsParams = {
  categoryId?: string;
  q?: string;
};

export type ListMyOrdersParams = {
  status?: ListMyOrdersStatus;
};

export type ListMyOrdersStatus =
  (typeof ListMyOrdersStatus)[keyof typeof ListMyOrdersStatus];

export const ListMyOrdersStatus = {
  all: "all",
  wait: "wait",
  accept: "accept",
  reject: "reject",
} as const;

export type ListMyDepositsParams = {
  status?: ListMyDepositsStatus;
  method?: string;
};

export type ListMyDepositsStatus =
  (typeof ListMyDepositsStatus)[keyof typeof ListMyDepositsStatus];

export const ListMyDepositsStatus = {
  all: "all",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;