# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Current user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.string(),
  telegramId: zod.string(),
  username: zod.string(),
  balanceUsd: zod.number(),
  balanceSyp: zod.number(),
  role: zod.enum(["user", "agent", "admin"]),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  image: zod.string(),
  order: zod.number(),
  active: zod.boolean(),
  productCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List products
 */
export const ListProductsQueryParams = zod.object({
  categoryId: zod.coerce.string().optional(),
  q: zod.coerce.string().optional(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  categoryId: zod.string(),
  categoryName: zod.string(),
  image: zod.string(),
  priceUsd: zod.number(),
  priceSyp: zod.number(),
  productType: zod.enum(["amount", "package"]),
  available: zod.boolean(),
  minQty: zod.number().optional(),
  maxQty: zod.number().optional(),
  description: zod.string().optional(),
  featured: zod.boolean(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Featured products for home slider
 */
export const ListFeaturedProductsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  categoryId: zod.string(),
  categoryName: zod.string(),
  image: zod.string(),
  priceUsd: zod.number(),
  priceSyp: zod.number(),
  productType: zod.enum(["amount", "package"]),
  available: zod.boolean(),
  minQty: zod.number().optional(),
  maxQty: zod.number().optional(),
  description: zod.string().optional(),
  featured: zod.boolean(),
});
export const ListFeaturedProductsResponse = zod.array(
  ListFeaturedProductsResponseItem,
);

export const GetProductParams = zod.object({
  id: zod.coerce.string(),
});

export const GetProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  categoryId: zod.string(),
  categoryName: zod.string(),
  image: zod.string(),
  priceUsd: zod.number(),
  priceSyp: zod.number(),
  productType: zod.enum(["amount", "package"]),
  available: zod.boolean(),
  minQty: zod.number().optional(),
  maxQty: zod.number().optional(),
  description: zod.string().optional(),
  featured: zod.boolean(),
});

/**
 * @summary News ticker entries
 */
export const ListNewsResponseItem = zod.object({
  id: zod.string(),
  content: zod.string(),
  type: zod.enum(["general", "offer", "alert", "new_service"]),
});
export const ListNewsResponse = zod.array(ListNewsResponseItem);

/**
 * @summary Home banner slider
 */
export const ListBannersResponseItem = zod.object({
  id: zod.string(),
  image: zod.string(),
  title: zod.string(),
  link: zod.string().optional(),
});
export const ListBannersResponse = zod.array(ListBannersResponseItem);

/**
 * @summary Available deposit payment methods
 */
export const ListPaymentMethodsResponseItem = zod.object({
  id: zod.string(),
  code: zod.enum([
    "sham_cash",
    "binance_pay",
    "syriatel_cash",
    "mtn_cash",
    "usdt_auto",
  ]),
  name: zod.string(),
  subtitle: zod.string(),
  instructions: zod.string().optional(),
  walletAddress: zod.string().optional(),
  qrImage: zod.string().optional(),
  minAmount: zod.number(),
  active: zod.boolean(),
});
export const ListPaymentMethodsResponse = zod.array(
  ListPaymentMethodsResponseItem,
);

/**
 * @summary Support / social links
 */
export const ListSocialLinksResponseItem = zod.object({
  id: zod.string(),
  platform: zod.string(),
  url: zod.string(),
  label: zod.string(),
});
export const ListSocialLinksResponse = zod.array(ListSocialLinksResponseItem);

/**
 * @summary Current user orders
 */
export const ListMyOrdersQueryParams = zod.object({
  status: zod.enum(["all", "wait", "accept", "reject"]).optional(),
});

export const ListMyOrdersResponseItem = zod.object({
  id: zod.string(),
  orderNumber: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productImage: zod.string().optional(),
  quantity: zod.number(),
  userIdentifier: zod.string().optional(),
  totalUsd: zod.number(),
  totalSyp: zod.number(),
  status: zod.enum(["wait", "accept", "reject"]),
  createdAt: zod.string(),
});
export const ListMyOrdersResponse = zod.array(ListMyOrdersResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  productId: zod.string(),
  quantity: zod.number(),
  userIdentifier: zod.string().optional(),
});

export const CreateOrderResponse = zod.object({
  id: zod.string(),
  orderNumber: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productImage: zod.string().optional(),
  quantity: zod.number(),
  userIdentifier: zod.string().optional(),
  totalUsd: zod.number(),
  totalSyp: zod.number(),
  status: zod.enum(["wait", "accept", "reject"]),
  createdAt: zod.string(),
});

/**
 * @summary Aggregated order stats for current user
 */
export const GetOrdersSummaryResponse = zod.object({
  totalAcceptedUsd: zod.number(),
  waitCount: zod.number(),
  acceptCount: zod.number(),
  totalCount: zod.number(),
});

export const GetOrderParams = zod.object({
  id: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  id: zod.string(),
  orderNumber: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productImage: zod.string().optional(),
  quantity: zod.number(),
  userIdentifier: zod.string().optional(),
  totalUsd: zod.number(),
  totalSyp: zod.number(),
  status: zod.enum(["wait", "accept", "reject"]),
  createdAt: zod.string(),
});

/**
 * @summary Current user deposits
 */
export const ListMyDepositsQueryParams = zod.object({
  status: zod.enum(["all", "pending", "approved", "rejected"]).optional(),
  method: zod.coerce.string().optional(),
});

export const ListMyDepositsResponseItem = zod.object({
  id: zod.string(),
  amountUsd: zod.number(),
  amountSyp: zod.number().optional(),
  currency: zod.enum(["USD", "SYP"]),
  method: zod.string(),
  methodLabel: zod.string(),
  transactionId: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.string(),
});
export const ListMyDepositsResponse = zod.array(ListMyDepositsResponseItem);

/**
 * @summary Submit a manual deposit
 */
export const CreateDepositBody = zod.object({
  method: zod.string(),
  currency: zod.enum(["USD", "SYP"]),
  amount: zod.number(),
  transactionId: zod.string(),
});

export const CreateDepositResponse = zod.object({
  id: zod.string(),
  amountUsd: zod.number(),
  amountSyp: zod.number().optional(),
  currency: zod.enum(["USD", "SYP"]),
  method: zod.string(),
  methodLabel: zod.string(),
  transactionId: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.string(),
});

/**
 * @summary Aggregated deposit stats
 */
export const GetDepositsSummaryResponse = zod.object({
  totalApprovedUsd: zod.number(),
  pendingCount: zod.number(),
  approvedCount: zod.number(),
  totalCount: zod.number(),
});