# lib/api-zod/src/generated/types/depositsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */

export interface DepositsSummary {
  totalApprovedUsd: number;
  pendingCount: number;
  approvedCount: number;
  totalCount: number;
}