# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  orderNumber: string;
  productId: string;
  productName: string;
  productImage?: string;
  quantity: number;
  userIdentifier?: string;
  totalUsd: number;
  totalSyp: number;
  status: OrderStatus;
  createdAt: string;
}