# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethodCode } from "./paymentMethodCode";

export interface PaymentMethod {
  id: string;
  code: PaymentMethodCode;
  name: string;
  subtitle: string;
  instructions?: string;
  walletAddress?: string;
  qrImage?: string;
  minAmount: number;
  active: boolean;
}