# lib/api-zod/src/generated/types/socialLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XPayStore API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLink {
  id: string;
  platform: string;
  url: string;
  label: string;
}